refactor: move the exporter entry into an internal package
The run logic moves out of the command-line binary into an internal package that the binary now calls, so tests can drive the same entry in-process. Each run builds its own flag application and metric registry instead of mutating process-global state, registering the same collectors and instrumented metrics handler the global registry used to provide. Installing the configured logger as the process default became an explicit option that only the binary enables, and callers can pass a hook reporting the bound listener addresses, which lets tests listen on an ephemeral port. The Windows service passes the process arguments to the entry explicitly, keeping the flags of an installed service intact.

The served metrics, the help output and the root page were verified to be byte-identical before and after the change.

Part of utkuozdemir/nvidia_gpu_exporter#422.
